pengtai | standardise scraper for google news
